requirements.txt
flask>=3.0.0
torch>=2.1.0
torchvision>=0.16.0
torchxrayvision>=1.2.0
opencv-python>=4.8.0
pillow>=10.0.0
pydicom>=2.4.0
werkzeug>=3.0.0
numpy>=1.24.0,<2.0.0
